Refactor installer default.nix to add support for autologin user
This commit refactors the `installer/default.nix` file to add support for autologin user functionality. The `services.getty.autologinUser` configuration option is set to the value of the `username` variable. This change enables automatic login for the specified user on the installed system.

The code changes include modifying the `dependencies` list to include the necessary paths for building the installer, as well as updating the `environment.systemPackages` list to include a new shell script that installs NixOS unattended using the `disko-install` command with the specified flake.

These changes improve the installer functionality by allowing the user to specify an autologin user during the installation process.